Move checkValueSatisfiesCondition to @tupaia/utils
