+++ packages/aggregator/src/analytics/aggregateAnalytics/aggregations/entityAggregations.js
@@ -3,6 +3,8 @@
  * Copyright (c) 2019 Beyond Essential Systems Pty Ltd
  */
 
+import { checkValueSatisfiesCondition } from '@tupaia/utils';
+
 /**
  * Combines analytics together across org units
  *
@@ -11,9 +13,11 @@
  * @param {Function} baseValueMapper
  * @param {Function} keyMapper
  */
-const valuePerOrgGroup = (analytics, aggregationConfig, baseValueMapper, keyMapper) => {
-  const { orgUnitMap = {}, valueToMatch } = aggregationConfig;
-  const valueMapper = valueToMatch ? createConfiguredValueMapper(valueToMatch) : baseValueMapper;
+const valuePerOrgGroup = (analytics, aggregationConfig, valueMapper, keyMapper) => {
+  const { orgUnitMap = {}, valueToMatch, condition } = aggregationConfig;
+  let filterValueMapper = () => true;
+  if (valueToMatch) filterValueMapper = createFilterValueMapper(valueToMatch);
+  if (condition) filterValueMapper = createFilterValueMapper(condition);
 
   const analyticsByKey = {};
   analytics.forEach(analytic => {
@@ -21,7 +25,7 @@ const valuePerOrgGroup = (analytics, aggregationConfig, baseValueMapper, keyMapp
       (orgUnitMap[analytic.organisationUnit] && orgUnitMap[analytic.organisationUnit].code) ||
       analytic.organisationUnit;
     const key = keyMapper(analytic, organisationUnit);
-    const value = valueMapper(analytic.value);
+    const value = filterValueMapper(analytic.value) ? 0 : valueMapper(analytic.value);
 
     // If there are no matching response elements already being returned, add it
     if (!analyticsByKey[key]) {
@@ -70,8 +74,7 @@ export const countPerOrgGroup = (analytics, aggregationConfig) =>
 export const countPerPeriodPerOrgGroup = (analytics, aggregationConfig) =>
   valuePerOrgGroup(analytics, aggregationConfig, countValueMapper, perPeriodPerOrgGroupKey);
 
-const createConfiguredValueMapper = valueToMatch =>
-  valueToMatch === '*' ? () => 1 : value => (value === valueToMatch ? 1 : 0);
+const createFilterValueMapper = condition => val => checkValueSatisfiesCondition(val, condition);
 
 const sumValueMapper = value => value;
