+++ packages/utils/src/index.js
@@ -29,3 +29,4 @@ export { respond } from './respond';
 export { singularise, stripFromString } from './string';
 export * from './validation';
 export { WorkBookParser } from './WorkBookParser';
+export { checkValueSatisfiesCondition } from './checkValueSatisfiesCondition';
